Add QR code flag to goshy
This commit adds an additional flag for displaying an QR code as the
respone of running goshy. It requires the qrencode package.

One could check if the depedency exists, but any OS should be able to
report this themselves. So this bloat was not added.

Sadly `qrencode -t ASCII` was not working for me. It printed a not
scannable QR code. Thus an image is opened.